+++ packages/client/src/index.ts
@@ -5,5 +5,5 @@
 
 export type { IConnectionProvider, MonacoLanguageClientOptions } from './monaco-language-client.js';
 export * from './monaco-language-client.js';
-export type { InitializeServiceConfig } from './monaco-vscode-api-services.js';
+export type { InitializeServiceConfig, MonacoEnvironmentEnhanced } from './monaco-vscode-api-services.js';
 export * from './monaco-vscode-api-services.js';